Refactor job_finder, driver and job_emailer to: finder, jobfinder, and emailer
